automatically cleanup old defunct AMIs
automatically clear our old defunct AMIs that are no longer being
used. this running once a day should give us plenty of room for
rollbacks as those can be executed quickly <10 minutes.

this is not only a cost saving measure, but also makes it much easier
to reason about the state of the world.

Signed-off-by: Cynthia <cynthia@coan.dev>